Addresses feedback from PR.
