+++ src/vs/monaco.d.ts
@@ -1396,6 +1396,9 @@ declare namespace monaco.editor {
 		readonly endColumn: number;
 	}
 
+	/**
+	 * Provides tokenization related functionality of the text model.
+	*/
 	export interface ITokenizationTextModelPart {
 		/**
 		 * Get the word under or besides `position`.
@@ -1937,6 +1940,18 @@ declare namespace monaco.editor {
 		 * Get the language associated with this model.
 		 */
 		getLanguageId(): string;
+		/**
+		 * Get the word under or besides `position`.
+		 * @param position The position to look for a word.
+		 * @return The word under or besides `position`. Might be null.
+		 */
+		getWordAtPosition(position: IPosition): IWordAtPosition | null;
+		/**
+		 * Get the word under or besides `position` trimmed to `position`.column
+		 * @param position The position to look for a word.
+		 * @return The word under or besides `position`. Will never be null.
+		 */
+		getWordUntilPosition(position: IPosition): IWordAtPosition;
 		/**
 		 * Perform a minimum amount of operations, in order to transform the decorations
 		 * identified by `oldDecorations` to the decorations described by `newDecorations`
